+++ CHANGES.rst
@@ -14,6 +14,7 @@ Unreleased
 ----------
  * **[FEATURE]** Add methods `follow_tag` and `unfollow_tag` to `User` object. This can be used to follow tags like "cats" on Imgur.
   * **[FEATURE]** Add method `get_votes` on `Gallery_image` and `Gallery_album`, to get the upvotes and downvotes.
+ * **[FEATURE]** Update `search_gallery` to support advanced querying, sort on time/viral, limit on time window and add pagination.
  * **[FEATURE]** Add exponential backoff logic with jitter to requests that fail on
    issues, which PyImgur believes are transient issues. Backoff starts at 2 seconds.
  * **[FEATURE]** It is no longer neccessary to call get_refresh_token() to get